+++ config/config.json
@@ -2,9 +2,9 @@
   "development": {
     "username": "root",
 
-    "password": "",
+    
 
-    "password": "Password1",
+    "password": "FrankY242424!",
 
     "database": "bid_items",
     "host": "127.0.0.1",
