+++ public/js/index.js
@@ -1,6 +1,8 @@
 // $(document).ready(() => {
 
 // Return to top button function
+const emailInput = $("input#email-input");
+const passwordInput = $("input#password-input");
 homebutton = document.getElementById("homeBtn");
 
 
@@ -48,5 +50,36 @@ $("#submitUpload").on("click", (event) => {
       console.log(err);
     }
   });
+  function loginUser(email, password) {
+    $.post("/api/login", {
+      email: email,
+      password: password
+    })
+      .then(() => {
+        window.location.replace("/index");
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  }
+    
+    
+  $("#siteLog").on ("submit"), event => {
+    event.preventDefault();
+    const userData = {
+      email: emailInput.val().trim(),
+      password: passwordInput.val().trim()
+    };
+    
+    if (!userData.email || !userData.password) {
+      return;
+    }
+    
+      
+    loginUser(userData.email, userData.password);
+    emailInput.val("");
+    passwordInput.val("");
+  };
+    
 });
 // });
