+++ public/js/login.js
@@ -1,40 +1,40 @@
-$(document).ready(() =>{
+// $(document).ready(() =>{
 
-  const loginForm = $("form.login");
-  const emailInput = $("input#email-input");
-  const passwordInput = $("input#password-input");
-  function loginUser(email, password) {
-    $.post("/api/login", {
-      email: email,
-      password: password
-    })
-      .then(() => {
-        window.location.replace("/members");
-      })
-      .catch((err) => {
-        console.log(err);
-      });
-  }
+  
+//   const emailInput = $("input#email-input");
+//   const passwordInput = $("input#password-input");
+//   function loginUser(email, password) {
+//     $.post("/api/login", {
+//       email: email,
+//       password: password
+//     })
+//       .then(() => {
+//         window.location.replace("/index");
+//       })
+//       .catch((err) => {
+//         console.log(err);
+//       });
+//   }
     
     
-  loginForm.on ("submit"), event => {
-    event.preventDefault();
-    const userData = {
-      email: emailInput.val().trim(),
-      password: passwordInput.val().trim()
-    };
+//   $("#siteLog").on ("submit"), event => {
+//     event.preventDefault();
+//     const userData = {
+//       email: emailInput.val().trim(),
+//       password: passwordInput.val().trim()
+//     };
     
-    if (!userData.email || !userData.password) {
-      return;
-    }
+//     if (!userData.email || !userData.password) {
+//       return;
+//     }
     
       
-    loginUser(userData.email, userData.password);
-    emailInput.val("");
-    passwordInput.val("");
-  };
+//     loginUser(userData.email, userData.password);
+//     emailInput.val("");
+//     passwordInput.val("");
+//   };
     
     
   
-});
+// });
