+++ views/layouts/main.handlebars
@@ -41,7 +41,10 @@ function openForm() {
 function closeForm() {
   document.getElementById("myForm").style.display = "none";
 }
+
+
 </script>
+<script src ="/js/login.js"></script>
 <script src="/js/index.js"></script>
 </body>
