+++ views/partials/nav/nav-block.handlebars
@@ -21,7 +21,7 @@
 </div>
 
 <div class="form-popup" id="myForm">
-  <form action="/action_page.php" class="form-container">
+  <form  class="form-container">
     <h1>Login</h1>
 
     <label for="email"><b>Email</b></label>
